Update ViewImplementation.cpp
In The Core::System::make_directory function is called with the path to the downloads directory and a permission mode of 0755 (read/write/execute for owner, read/execute for group and others). I believe that is correct. 

This approach should prevent the "open: No such file or directory (errno=2)" from happening, bc the downloads folder will be created